fix(editor3): Move the selection iterator in it's on file
